tests/wpt/web-platform-tests/css/compositing/parsing/mix-blend-mode-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Compositing and Blending Level 1: parsing mix-blend-mode with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/compositing-1/#propdef-mix-blend-mode">
<meta name="assert" content="mix-blend-mode supports the full grammar '<blend-mode>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("mix-blend-mode", "normal");
test_valid_value("mix-blend-mode", "multiply");
test_valid_value("mix-blend-mode", "screen");
test_valid_value("mix-blend-mode", "overlay");
test_valid_value("mix-blend-mode", "darken");
test_valid_value("mix-blend-mode", "lighten");
test_valid_value("mix-blend-mode", "color-dodge");
test_valid_value("mix-blend-mode", "color-burn");
test_valid_value("mix-blend-mode", "hard-light");
test_valid_value("mix-blend-mode", "soft-light");
test_valid_value("mix-blend-mode", "difference");
test_valid_value("mix-blend-mode", "exclusion");
test_valid_value("mix-blend-mode", "hue");
test_valid_value("mix-blend-mode", "saturation");
test_valid_value("mix-blend-mode", "color");
test_valid_value("mix-blend-mode", "luminosity");
</script>
</body>
</html>
